requirements.txt
opencv-python
ultralytics
torch
torchvision
torchaudio
flask
flask_socketio
python-dotenv